删除Swift protocol optional method 讨论
Chris Latter宣布不支持

We discussed this extensively in the Swift 3 timeframe, and I believe
that the consensus was “no”: optional requirements are considered an
Objective-C compatibility feature.

I don’t recall all of the discussion, but a major one would have to be
that it is highly redundant with other language features we already
have, and makes the design space of Swift APIs more complicated.  If
Objective-C didn’t already have them, we would never consider adding
them to our protocol model.

-Chris